Parsing of optional INFO fields
Reserved fields, according to the VCF 4.2 specification, section 1.4.1 part 8, are parsed
and set to several pointer fields. This allows nil values if the fields can't be parsed
successfully.

These fields are meant as helpers, since the specification defines they are dependent on
the metadata on the header, which is not used.